Hash tag implementation ,page scroll issue solved
1. After implementing hash tags, the followers list was scrolling when a
follower was clicked as the <a> had followers name as "id" attribute
which we are adding to url as hash when follower is clicked. This is a
default behavior.

2. To solve this, "id" attribute of the <a> tag is prefixed by "id-" so
that there is no element on the page that has "id" property same as that
of the hash that we are appending to the URL